hps-system/agente-ia/requirements.txt
# =============================================================================
# DEPENDENCIAS PARA EL AGENTE IA HPS
# =============================================================================

# FastAPI y servidor
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
websockets>=11.0.0

# OpenAI
openai>=1.3.0

# Base de datos
asyncpg>=0.29.0
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0
alembic>=1.12.0

# Email
emails>=0.6.0
jinja2>=3.1.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
click>=8.1.0
python-jose>=3.3.0

# Autenticación y seguridad
passlib>=1.7.4
bcrypt>=4.0.0
email-validator>=2.0.0

# HTTP y requests
httpx>=0.24.0
requests>=2.31.0

# Logging
structlog>=23.1.0

# Desarrollo y testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
